Refactor security into service
